+++ __pathperror.c
@@ -13,9 +13,9 @@ __attribute__((visibility("hidden")))
 void __pathperror(const char *path, const char *s)
 {
 	if ((errno != EPERM) && (errno != EACCES)) {
-		__info("%s: %s: %m\n", path, s);
+		__info("%s: %s: %s: %m\n", getrootdir(), path, s);
 		return;
 	}
 
-	__note_or_fatal("%s: %s: %m\n", path, s);
+	__note_or_fatal("%s: %s: %s: %m\n", getrootdir(), path, s);
 }