+++ __pathperror2.c
@@ -13,9 +13,11 @@ __attribute__((visibility("hidden")))
 void __pathperror2(const char *oldpath, const char *newpath, const char *s)
 {
 	if ((errno != EPERM) && (errno != EACCES)) {
-		__info("%s: %s: %s: %m\n", oldpath, newpath, s);
+		__info("%s: %s: %s: %s: %m\n", getrootdir(), oldpath, newpath,
+		       s);
 		return;
 	}
 
-	__note_or_fatal("%s: %s: %s: %m\n", oldpath, newpath, s);
+	__note_or_fatal("%s: %s: %s: %s: %m\n", getrootdir(), oldpath, newpath,
+			s);
 }